requirements.txt
torch==2.1.0
torchvision==0.16.0
matplotlib==3.7.1
pytest==7.4.0
numpy==1.24.3
pillow>=9.0.0